ABI: Move snapshots/environments integration to Partout (#1815)
Move what was formerly observed by TunnelManager to Tunnel and
SimpleConnectionDaemon in Partout, following
https://github.com/partout-io/partout/pull/400 and
https://github.com/partout-io/partout/pull/401. The timer is not
necessary in TunnelManager because the Tunnel snapshots are now
thorough.

Let the snapshots IPC happen internally between PartoutVpnServiceRuntime
and PartoutTunnel, thus making the status handler binding redundant and
the TunnelManager events flow naturally.

Add data count to the Android app to showcase the new metadata.

+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/abi/models/ProfileTransfer.kt
@@ -31,10 +31,10 @@ import kotlinx.serialization.Contextual
 data class ProfileTransfer (
 
     @SerialName(value = "received")
-    val received: kotlin.Int = 0,
+    val received: kotlin.Long = 0L,
 
     @SerialName(value = "sent")
-    val sent: kotlin.Int = 0
+    val sent: kotlin.Long = 0L
 
 ) {
